Added corrections for issue #2 double changed to int

+++ src/main/java/org/example/Solver.java
@@ -3,7 +3,7 @@ package org.example;
 public class Solver {
     double d;
 
-    public double resultOfDiscriminant(int a, int b, int c) {
+    public int resultOfDiscriminant(int a, int b, int c) {
         int result;
         d = b * b - 4 * a * c;
         if (d < 0) {
